pyproject.toml
[project]
name = "gui"
version = "0.1.0"
description = "Copier template for generic Python 3 project with tooling configuration for testing, linting and IDE support"
authors = [{ name = "Li Lin", email = "guru.lin@gmail.com" }]
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "bootstrap-flask>=2.4.1",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv >= 1.0.1",
    "google-cloud-logging>=3.11.3",
]
[dependency-groups]
dev = [
    "pytest >= 8.3.0",
    "pytest-dotenv >= 0.5.2",
    "pytest-mock >= 3.14",
    "pytest-cov >= 6.0",
    "pre-commit >= 4.0.1",
    "ruff >= 0.8.2",
    "pyright >= 1.1.139",
]

[tool.uv]
managed = true

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]
pythonpath = ["."]
filterwarnings = [
    "ignore::DeprecationWarning",
]
env_files = [".env"]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
]

# see ruff configuration options at
# https://docs.astral.sh/ruff/configuration/

[tool.ruff]
line-length = 90
indent-width = 4
exclude = [".git", "__pycache__", "venv", ".venv"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "C",
    "I",
    "A",
] # Enable checks for PEP8, imports, and other standard rules
ignore = ["E203", "E266"]


[tool.ruff.lint.per-file-ignores]
"migrations/*.py" = ["E501"] # Ignore line too long for auto-generated migration files


[tool.aerich]
tortoise_orm = "main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
